fix: fixed issues wit nltk and logs printing (#129)
Simplified logic responsible for downloading nltk data. Try/except is not needed because when package is already downloaded,
it will be not downloaded again so just `nltk.download()` is enough.

Also changed 'punkt' to 'pubkt_tab' following to error I saw in the console.

Besides, fixed small issue with one log that was printed to stdout, but should be printend to stderr